recursively apply messages to extra from throw
extraFromError does its own work recursively but it's
not responsible for processing error messages. Attach
these in threw() like normal, but do so recursively,
walking the error and extra objects simultaneously.

This fundamentally requires that the structures of the
provided error and extra objects match each other 1:1.
There's no getting around that because otherwise we
have no way to associate errors and extras. This means
we're hard-coding expectations about when extraFromError
provides `cause` and `errors` values.