recursively call extraFromError on cause & errors
This doesn't need to be guarded as heavily as cleanYamlObject
because extraFromError is already intended to handle all kinds
of error shapes (including non-objects). The basic behavior is
the same though.

NOTE: Because we treat values stored on cause & errors the
same way as anything else passed to extraFromError, this DOES
cause values that aren't or don't look like Error objects to
get passed through. This has particular impact on plan-related
errors (see #printResult) which set some diagnostic details on
the constructed error object.